+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v1.11.2
+
+- Check for the existence of a viewconf before creating the link
+
+v1.11.1
+
+- Switch to using networkidle0
+
+v1.11.0
+
+- Link unfurling endpoints /link and /thumbnail
+
 v1.10.2
 
 - Added project to the admin interface